+++ src/tubespbo/Model/Driver.java
@@ -15,7 +15,7 @@ public class Driver extends User {
     private String vehicle_type;
     private String vehicle_plate;
 
-    public Driver(int driver_id, String driver_phonNum, String vehicle_name, String vehicle_type, String vehicle_plate, int user_id, String user_name, String user_pass, String user_role, double user_wallet) {
+    public Driver(int driver_id, String driver_phonNum, String vehicle_name, String vehicle_type, String vehicle_plate, int user_id, String user_name, String user_pass, String user_role, float user_wallet) {
         super(user_id, user_name, user_pass, user_role, user_wallet);
         this.driver_id = driver_id;
         this.driver_phonNum = driver_phonNum;
